backend/requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Werkzeug==3.0.1

# Database
PyMySQL==1.1.0
SQLAlchemy==2.0.25
cryptography==41.0.7  # Required for PyMySQL

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2

# Environment variables
python-dotenv==1.0.0

# Development tools
black==23.12.1  # Code formatter
flake8==7.0.0   # Code linter
pytest==7.4.4   # Testing

# Production server
gunicorn==21.2.0

# Utilities
python-dateutil==2.8.2
six==1.16.0 